pyproject.toml
[project]
name = "youtube-summarizer-agent"
version = "2.0.0"
description = "Agente LangGraph que usa Google Gemini para resumir vídeos do YouTube"
authors = [
    {name = "YouTube Summarizer Team", email = "dev@youtube-summarizer.com"},
]
dependencies = [
    "langchain>=0.1.6",
    "langgraph>=0.0.26",
    "google-generativeai>=0.8.3",
    "youtube-transcript-api>=0.6.2",
    "python-dotenv>=1.0.0",
    "streamlit>=1.31.0",
    "validators>=0.22.0",
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/youtube-summarizer/youtube-summarizer-agent"
Repository = "https://github.com/youtube-summarizer/youtube-summarizer-agent"
Documentation = "https://github.com/youtube-summarizer/youtube-summarizer-agent#readme"
Issues = "https://github.com/youtube-summarizer/youtube-summarizer-agent/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
youtube-summarizer = "youtube_summarizer_agent:main"
youtube-summarizer-web = "streamlit_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["youtube_summarizer_agent", "config"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
] 